updated and added debug/comments to files

+++ keras_frcnn/vgg.py
@@ -28,7 +28,9 @@ def get_weight_path():
 
 def get_img_output_length(width, height):
     def get_output_length(input_length):
-        return input_length/16
+        # print('vgg dim resize div:', input_length/16)
+        # return input_length/16
+        return int(input_length / 16)
 
     return get_output_length(width), get_output_length(height)
